apps/worker-gateway/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx>=0.28,<1.0
pydantic-settings>=2.7,<3.0
pyyaml>=6.0,<7.0
asyncpg>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
